create po_up_window super class
